+++ src/baobab_collection_core/exceptions/__init__.py
@@ -2,9 +2,12 @@
 
 from .baobab_collection_core_exception import BaobabCollectionCoreException
 from .card_not_found_exception import CardNotFoundException
+from .container_cycle_exception import ContainerCycleException
+from .container_not_found_exception import ContainerNotFoundException
 from .duplicate_card_exception import DuplicateCardException
 from .duplicate_user_exception import DuplicateUserException
 from .invalid_card_exception import InvalidCardException
+from .invalid_container_exception import InvalidContainerException
 from .invalid_physical_copy_exception import InvalidPhysicalCopyException
 from .invalid_user_exception import InvalidUserException
 from .physical_copy_not_found_exception import PhysicalCopyNotFoundException
@@ -14,9 +17,12 @@ from .validation_exception import ValidationException
 __all__: list[str] = [
     "BaobabCollectionCoreException",
     "CardNotFoundException",
+    "ContainerCycleException",
+    "ContainerNotFoundException",
     "DuplicateCardException",
     "DuplicateUserException",
     "InvalidCardException",
+    "InvalidContainerException",
     "InvalidPhysicalCopyException",
     "InvalidUserException",
     "PhysicalCopyNotFoundException",